requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
black==24.2.0
cbor2==5.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cytoolz==0.12.3
dataclassy==0.11.1
debugpy==1.8.1
decorator==5.1.1
eip712==0.2.4
eth-account==0.10.0
eth-brownie==1.20.2
eth-event==1.2.5
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.0.0
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
frozenlist==1.4.1
hexbytes==0.3.1
hypothesis==6.27.3
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lazy-object-proxy==1.10.0
lru-dict==1.2.0
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parsimonious==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-solc-ast==1.2.10
py-solc-x==1.1.1
py-solidity-ast==0.1.5
pyarrow==15.0.2
pycryptodome==3.20.0
Pygments==2.17.2
pygments-lexer-solidity==0.7.0
pytest==6.2.5
pytest-forked==1.6.0
pytest-xdist==1.34.0
python-dateutil==2.9.0.post0
python-dotenv==0.16.0
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vvm==0.1.0
vyper==0.3.10
wcwidth==0.2.13
web3==6.15.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
